% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt_dpapi_pw.R
\name{decrypt_dpapi_pw}
\alias{decrypt_dpapi_pw}
\title{Decrypt passwords encrypted with  the Microsoft Data Protection API}
\usage{
decrypt_dpapi_pw(file)
}
\arguments{
\item{file}{File that holds a password encrypted using DPAPI}
}
\description{
Decrypt passwords encrypted with  the Microsoft Data Protection API
}

