% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt_gk_pw.R
\name{decrypt_gk_pw}
\alias{decrypt_gk_pw}
\title{Get a password from Gnome Keyring using secret-tool}
\usage{
decrypt_gk_pw(key_value_pairs)
}
\arguments{
\item{key_value_pairs}{A string of key value pairs as expected by secret-tool (see example)}
}
\value{
A decrypted password as a string.
}
\description{
Get a password from Gnome Keyring using secret-tool
}
\examples{
\dontrun{
# returns the password for the key-value pair key: MYSCHEMA_MYDB
decrypt_gk_pw("key MYSCHEMA_MYDB")
}
}

