% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt_kc_pw.R
\name{decrypt_kc_pw}
\alias{decrypt_kc_pw}
\title{Get a password from OSX Keyring using 'security' cli}
\usage{
decrypt_kc_pw(key_value_pairs)
}
\arguments{
\item{label}{Keychain password label}
}
\value{
An decrypted password as an invisible string.  Invisible means that
the password won't be displayed in the console, but can be assigned to a
variable or used inline.
}
\description{
Get a password from OSX Keyring using 'security' cli
}
\examples{
\dontrun{
# First store a password in Keychain
# now return the password above to the R environment
x <- decrypt_kc_pw("label")

# function is best used in a connection string command:
ch <- odbcConnect("some dsn", uid = "user1", pwd = decrypt_kc_pw("mydb_myuser"))
}
}

