% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decryptGnomeKeyring.R
\name{decryptGnomeKeyring}
\alias{decryptGnomeKeyring}
\title{Get a password from Gnome Keyring using secret-tool}
\usage{
decryptGnomeKeyring(key_value_pairs)
}
\arguments{
\item{key_value_pairs}{A list of key value pairs as expected by secret-tool (see example)}
}
\value{
A decrypted password as a string.
}
\description{
Get a password from Gnome Keyring using secret-tool
}
\examples{
\dontrun{
# returns the password for the key-value pair key: MYSCHEMA_MYDB
decryptGnomeKeyring(key MYSCHEMA_MYDB)
}
}

