% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt_dpapi_pw.R
\name{decrypt_dpapi_pw}
\alias{decrypt_dpapi_pw}
\title{Decrypt passwords encrypted with  the Microsoft Data Protection API}
\usage{
decrypt_dpapi_pw(file)
}
\arguments{
\item{file}{File that holds a password encrypted using DPAPI}
}
\value{
An decrypted password as an invisible string.  Invisible means that
the password won't be displayed in the console, but can be assigned to a
variable or used inline.
}
\description{
Decrypt passwords encrypted with  the Microsoft Data Protection API
}
\details{
Requires Powershell to be installed and execution policy set to RemoteSigned.
}

